LiquidAnacron: Modularize API adapters
